requirements.txt
Django>=5.1.4
pandas>=2.2.3
numpy>=1.26.4
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openai>=1.10.0
